requirements.txt
pandas
numpy
matplotlib
typing
seaborn 
sklearn
imblearn
xgboost
pymongo
jupyter
ipykernel
python-dotenv
-e .